Move the backend DTOs and chat-export renderer out of AppModel.swift
The 33 file-private Codable wire structs move to Models/BackendResponses.swift
and ChatExportRenderer (with its save-panel accessory view) to its own file,
raised from private to internal so code outside AppModel.swift can keep using
them. This unpins the mobile-companion extension and the export entry points
from the monolith ahead of their own moves. No name collides module-wide;
both app targets build.

Behavior-identical refactor; project regenerated with xcodegen.
